+++ src/Catalog/CategorySummary.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * @author      Wizacha DevTeam <dev@wizacha.com>
+ * @copyright   Copyright (c) Wizacha
+ * @license     Proprietary
+ */
+
+declare(strict_types=1);
+
+namespace Wizaplace\SDK\Catalog;
+
+final class CategorySummary
+{
+    /** @var int */
+    private $id;
+    /** @var string */
+    private $name;
+    /** @var string */
+    private $slug;
+
+    /**
+     * @internal
+     */
+    public function __construct(array $data)
+    {
+        $this->id = $data['id'];
+        $this->name = $data['name'];
+        $this->slug = $data['slug'];
+    }
+
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function getSlug(): string
+    {
+        return $this->slug;
+    }
+}