+++ CHANGELOG.md
@@ -11,6 +11,13 @@
 
 </details>
 
+## 1.26.0
+
+### New features
+
+ - Added `\Wizaplace\SDK\Catalog\Category::getCategoryPath`
+ - new `\Wizaplace\SDK\Catalog\CategorySummary`
+
 ## 1.25.0
 
 ### New features